[FLINK-13632] In TypeSerializerUpgradeTestBase.UpgradeVerifier return matcher instead of data
Before, the test was using equals() to compare the expected data to the actual data. This does not work for types that don't have a proper equals() implementation. Now we return matchers and tests that use such types can return an appropriate matcher. Most tests can use the is() matcher, if the test data class has a proper equals() method.