[FLINK-27212][table-planner] Change printing of binary columns to use `x'ab3234f0'` format
Add a isPrinting() method to the CastRule.Context set to true used by RowDataToStringConverterImpl
which defines a different casting behaviour for BinaryToStringCastRule when printing binary columns, so
that we output columns of binary type as (for example): x'ab03f98e', which can easily be copy pasted as
a binary literal to another SQL query.

This closes #19507.